+++ packages/core/src/runtime/init.timingResolver.test.ts
@@ -1,5 +1,10 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { RuntimeTimelineLike } from "./types";
+import {
+  createMockTimeline,
+  installImmediateAnimationFrame,
+  resetRuntimeFixtureDom,
+  stubDuration,
+} from "./runtimeSeekFixture.test-helpers";
 
 const resolverSpy = vi.hoisted(() => ({ constructions: 0 }));
 const mediaSpy = vi.hoisted(() => ({ beforeSync: null as null | (() => void) }));
@@ -30,49 +35,13 @@ vi.mock("./media", async (importOriginal) => {
 
 const { initSandboxRuntimeModular } = await import("./init");
 
-function createMockTimeline(duration: number): RuntimeTimelineLike {
-  const state = { time: 0, paused: true };
-  return {
-    play: () => {
-      state.paused = false;
-    },
-    pause: () => {
-      state.paused = true;
-    },
-    seek: (time?: number) => (time === undefined ? state.time : (state.time = time)),
-    totalTime: (time?: number) => (time === undefined ? state.time : (state.time = time)),
-    time: () => state.time,
-    duration: () => duration,
-    add: () => {},
-    paused: (value?: boolean) =>
-      typeof value === "boolean" ? (state.paused = value) : state.paused,
-    timeScale: () => {},
-    set: () => {},
-    getChildren: () => [],
-  };
-}
-
-/** jsdom leaves `duration` as NaN; a writable getter also lets a test mimic the
- *  `el.load()` reset `syncRuntimeMedia` performs on the seek-retry path. */
-function stubDuration(el: HTMLMediaElement, initial: number): { set: (next: number) => void } {
-  let value = initial;
-  Object.defineProperty(el, "duration", { get: () => value, configurable: true });
-  return { set: (next: number) => (value = next) };
-}
-
 describe("runtime timing resolver scoping", () => {
   const originalRequestAnimationFrame = window.requestAnimationFrame;
   const originalCancelAnimationFrame = window.cancelAnimationFrame;
 
   beforeEach(() => {
-    document.body.innerHTML = "";
-    (globalThis as typeof globalThis & { CSS?: { escape?: (value: string) => string } }).CSS ??= {};
-    globalThis.CSS.escape ??= (value: string) => value;
-    window.requestAnimationFrame = ((callback: FrameRequestCallback) => {
-      callback(0);
-      return 1;
-    }) as typeof window.requestAnimationFrame;
-    window.cancelAnimationFrame = (() => {}) as typeof window.cancelAnimationFrame;
+    resetRuntimeFixtureDom();
+    installImmediateAnimationFrame();
     resolverSpy.constructions = 0;
     mediaSpy.beforeSync = null;
   });
